implement hash explicitely
Partial equality is implemented manually for the struct and only one
field is being used.
The derive hash trait works on all the fields. while at the sametime the
manual implementation only works on the address field.
This inconsistency between hash and eq can cause issues in things like
hashmaps.

Signed-off-by: Perma Alesheikh <me@prma.dev>