implement a consistent partial_eq and hash
Implementation of partial_eq and hash should always return a consistent
result. However, to have a consistent implementation, the fields that
would matter in the equality should be decided irrespective of the
compared case. Otherwise, when hashing, where only one of the cases are
present, we cannot predict which field should matter. And because of
that, our hashes won't match even though they are equal.

So with this implementation the hash-equality consistency is present,
however, the behavior of the equality is also changed slightly.
So in this case, if folder_a.kind and folder_b.kind have Some(T), but
their kinds don't match, they will return false for equality,
irrespective of their names.

Tests are also adjusted to reflect this change in behavior.

signed-off-by: Perma Alesheikh <me@prma.dev>
Signed-off-by: Perma Alesheikh <me@prma.dev>