rename new backend sync builder's to try_new
Since new usually indicates returning self, we should use another name
for this function. We can use 'build' for the name of the function,
however the function is recieving a builder, and making another builder,
and using 'build' can cause a bit of a confusion.

We can also use TryFrom, except that the function is neither under the
namespace of the input, nor is it under the namespace of the output.
So I just renamed it to try_new for now.

But I think there still is room for improvement on this one, TBH.

Signed-off-by: Perma Alesheikh <me@prma.dev>