remove anyhow dependency
Since anyhow is usually used as a dependency at the binary level; I have
removed the dependency and instead created a crate level trait for
type-erased errors, without depending on anyhow.

The end result should be compatible with anyhow and color-eyre crates.

The trait also contains a function `as_any` which allows for upcasting.
However a default implementation could not be made since the size of the
self is not declared at the trait level, instead the user should
implement the function, which is as simple as returning the "self".

Signed-off-by: Perma Alesheikh <me@prma.dev>